move slideout to own file, confirm messages

+++ public/js/project.js
@@ -2,7 +2,7 @@ $(document).ready(function () {
 
   let projectList = [];
   let completedList = [];
-  let inFocusProject = 0;
+  let inFocusIndex = 0;
   const userUrl = window.location.pathname;
 
   if(projectList.length < 1)
@@ -13,7 +13,7 @@ $(document).ready(function () {
     $.get('http://localhost:5000'+userUrl+'/liveprojects', function(projects){
       projects.forEach((project,index)=> {
 
-        if(inFocusProject === index){
+        if(inFocusIndex === index){
           $(".projects ul").append('<li class="list-project"><span id="bullet"><i class="fa fa-circle" aria-hidden="true"></i></span><span class="projectName">'+project.projectName+ '</span><form class="project-list-form" action="/liveprojects/' + project._id + '" method="POST"><span class="complete icon-wrapper-list"><i class="fa fa-check custom-icon-list" aria-hidden="true"><span class="fix-editor">&nbsp;</span></i></span></form><form class="project-list-form" action="/liveprojects/' + project._id + '" method="POST"><span class="delete icon-wrapper-list"><i class="fa fa-times custom-icon-list" aria-hidden="true"><span class="fix-editor">&nbsp;</span></i></span></form></li>');
           document.getElementById('taskInput').textContent = project.projectName;
 
@@ -29,24 +29,29 @@ $(document).ready(function () {
 
   //DELETE A LIVE PROJECT
   $("ul").on('click','.delete', function(event){
-    $(this).parent().fadeOut(500, function(){
 
-      const actionUrl = 'http://localhost:5000'+userUrl + $(this).attr('action');
-      const $itemToDelete = $(this).parent('.list-project');
-      const prevProject = $itemToDelete.prev();
-      const nextProject = $itemToDelete.next();
-      //console.log($itemToDelete);
-      $.ajax({
-        url: actionUrl,
-        type: 'DELETE',
-        itemToDelete: $itemToDelete,
-        success: function success(project){
-          //console.log(project);</form>
-          this.itemToDelete.remove();
-          deleteFromProjectList(project,prevProject,nextProject);
-        }
-      })
-    });
+    const confirm = window.confirm("Are you sure you want to delete this project and it's logged time?");
+
+    if (confirm){
+      $(this).parent().fadeOut(500, function(){
+
+        const actionUrl = 'http://localhost:5000'+userUrl + $(this).attr('action');
+        const $itemToDelete = $(this).parent('.list-project');
+        const prevProject = $itemToDelete.prev();
+        const nextProject = $itemToDelete.next();
+        //console.log($itemToDelete);
+        $.ajax({
+          url: actionUrl,
+          type: 'DELETE',
+          itemToDelete: $itemToDelete,
+          success: function success(project){
+            //console.log(project);</form>
+            this.itemToDelete.remove();
+            deleteFromProjectList(project,prevProject,nextProject);
+          }
+        })
+      });
+    }
     event.stopPropagation();
   });
 
@@ -54,26 +59,31 @@ $(document).ready(function () {
 
   //REMOVE A PROJECT FROM LIVE PROJECT LIST AND ADD TO COMPLETED PROJECTS
   $("ul").on('click','.complete', function(event){
-    $(this).parent().fadeOut(500, function(){
 
-      const actionUrl = 'http://localhost:5000'+userUrl + $(this).attr('action');
-      const $itemToUpdate = $(this).parent('.list-project');
-      const prevProject = $itemToUpdate.prev();
-      const nextProject = $itemToUpdate.next();
+    const confirm = window.confirm("Are you sure you want to mark this project as complete?");
 
-      $.ajax({
-        url: actionUrl,
-        data: {completed: true},
-        type: 'PUT',
-        itemToUpdate: $itemToUpdate,
-        success: function success(project){
-          //console.log(project);
-          this.itemToUpdate.remove();
-          deleteFromProjectList(project,prevProject,nextProject);
-        }
-      })
+    if (confirm){
+      $(this).parent().fadeOut(500, function(){
 
-    });
+        const actionUrl = 'http://localhost:5000'+userUrl + $(this).attr('action');
+        const $itemToUpdate = $(this).parent('.list-project');
+        const prevProject = $itemToUpdate.prev();
+        const nextProject = $itemToUpdate.next();
+
+        $.ajax({
+          url: actionUrl,
+          data: {completed: true},
+          type: 'PUT',
+          itemToUpdate: $itemToUpdate,
+          success: function success(project){
+            //console.log(project);
+            this.itemToUpdate.remove();
+            deleteFromProjectList(project,prevProject,nextProject);
+          }
+        })
+      });
+    }
+
     event.stopPropagation();
   });
 
@@ -90,8 +100,8 @@ $(document).ready(function () {
     $("#bullet i",this).removeClass('fa-circle-thin').addClass('fa-circle');
     document.getElementById('taskInput').textContent = $(".projectName",this).text();
 
-    inFocusProject = projectList.findIndex(x => x.projectName === $(".projectName",this).text());
-    //console.log(inFocusProject);
+    inFocusIndex = projectList.findIndex(x => x.projectName === $(".projectName",this).text());
+    //console.log(inFocusIndex);
   });
 
 
@@ -114,35 +124,44 @@ $(document).ready(function () {
     }
   });
 
-
+  //delete from the projectList array and select next Project as the inFocusIndex
   function deleteFromProjectList(project, prevProject, nextProject){
-    let arrIndex = projectList.findIndex(x => x._id === project._id);
+    //find the index of the deleted project in projectList
+    const deletedIndex = projectList.findIndex(x => x._id === project._id);
 
-    projectList.splice(arrIndex,1);
+    //remove the deleted project from the array
+    projectList.splice(deletedIndex,1);
 
+    //give the user the option to add new project if num projects < 5
     if(projectList.length < 5)
       $(".newProject").show();
 
-    if(inFocusProject > arrIndex)
-      inFocusProject --;
-
-    if(inFocusProject == arrIndex && projectList.length > 1){
-      $("#bullet i",nextProject).removeClass('fa-circle-thin').addClass('fa-circle');
-      document.getElementById('taskInput').textContent = $(".projectName",nextProject).text();
-      inFocusProject--;
+    //reasign selected project
+    //if inFocusIndex > deletedIndex => then the inFocusIndex is reduced by 1
+    //else if the projectList is now 0 then give user option to add projectList
+    //else if the deleted project is in focus and not the first project in the list, then the previous project becomes the infocus project
+    // else the deleted project is the first project in the list so the next project becomes the infocus project
+    if(inFocusIndex > deletedIndex)
+      inFocusIndex --;
+    else {
+      if(projectList<1){
+        document.getElementById('taskInput').textContent = "Add a project";
+      }else{
+        if(inFocusIndex == deletedIndex && deletedIndex > 0){
+          $("#bullet i",prevProject).removeClass('fa-circle-thin').addClass('fa-circle');
+          document.getElementById('taskInput').textContent = $(".projectName",prevProject).text();
+          inFocusIndex--;
+        }else{
+          $("#bullet i",nextProject).removeClass('fa-circle-thin').addClass('fa-circle');
+          document.getElementById('taskInput').textContent = $(".projectName",nextProject).text();
+        }
+      }
     }
+  };
 
-    if(inFocusProject == arrIndex && projectList.length < 1){
-      $("#bullet i",prevProject).removeClass('fa-circle-thin').addClass('fa-circle');
-      document.getElementById('taskInput').textContent = $(".projectName",prevProject).text();
-      inFocusProject--;
-    }
 
-    if(projectList<1){
-      document.getElementById('taskInput').textContent = "Add a project";
-    }
-  }
 
+  //Adds the newly added projects to the project list
   function addToLiveProjects(project){
     projectList.push(project);
     if(projectList.length === 5){
@@ -151,7 +170,7 @@ $(document).ready(function () {
 
     if(projectList.length === 1){
       document.getElementById('taskInput').textContent = projectList[0]['projectName'];
-      inFocusProject = 0;
+      inFocusIndex = 0;
       $(".projects ul").append('<li class="list-project"><span id="bullet"><i class="fa fa-circle" aria-hidden="true"></i></span><span class="projectName">'+project.projectName+ '</span><form class="project-list-form" action="/liveprojects/' + project._id + '" method="POST"><span class="complete icon-wrapper-list"><i class="fa fa-check custom-icon-list" aria-hidden="true"><span class="fix-editor">&nbsp;</span></i></span></form><form class="project-list-form" action="/liveprojects/' + project._id + '" method="POST"><span class="delete icon-wrapper-list"><i class="fa fa-times custom-icon-list" aria-hidden="true"><span class="fix-editor">&nbsp;</span></i></span></form></li>');
 
     }else{
